made few changes to snapd script

+++ snapd-purge-ubuntu.sh
@@ -2,7 +2,33 @@
 
 # Script for removing snapd from your Ubuntu system
 
-if dpkg --list | grep snapd > /dev/null; then
-	sudo systemctl disable snapd.socket
-	sudo apt-get remove snapd
+read -p "Are you sure you want to remove snap from your system? (yes/no) " yn
+echo
+
+if [[ $yn =~ ^[Yy]$ ]]
+then 
+	if dpkg --list | grep snapd > /dev/null; then
+		sudo systemctl disable snapd.socket
+		sudo apt-get remove snapd
+	else
+		echo "snap is not installed in your system"
+	fi
+else
+	exit
 fi
+
+echo
+read -p "Do you want to install flatpak in your system? (yes/no) " yn
+echo
+
+if [[ $yn =~ ^[Yy]$ ]]
+then 
+	if dpkg --list | grep flatpak > /dev/null; then
+		echo "flatpak is already installed in your system."
+		echo "Press any key to exit"
+	else
+		sudo apt-get install flatpak
+	fi
+fi
+
+
